separate out Class and call from station.js
